Add files via upload
After discussion, I implement an item length estimation method by eliminating the XX% longest outliers, for exmaple, 14%. More specifically, I think there are about 14% users who watch some certain video will leave and be inactive, so this 14% longest watch_interval for this single item is overestimated and should be eliminated. And this code implements this funtion and the results shows great improvement! The estimated video length become more rational and the ratio distribution become more rational.
PS. for  this method, using the 2to9 dataset, the watch_ratio more than 1 will be considered as the outlier data and be eliminated, for the item of total 108587 , there will be 551 estimated length =0 and 2771 estimated length >=3 hours
for the whole cutted  watch data of 9407057 entries,  after filtering the watch_ratio>=1 and <=0, the remain entries is 7514610 which is also rational.  shrink scale is about 80%
compare with the previous version 'get_watch_ratio_without_connection_without_dropduplicate.ipynb', which shrink the cutted orignal data from 23888514 to 20776898